requirements.txt

pillow
pygame
